# public/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "ProjectX Gateway API",
    "description": "ProjectX Gateway API Documentation and specification.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "produces": [
    "text/plain",
    "application/json",
    "text/json"
  ],
  "paths": {
...
    "/api/Position/searchOpen": {
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Searches for open positions based on the provided request.",
        "operationId": "Position_SearchOpenPositions",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing search criteria.",
            "schema": {
              "$ref": "#/definitions/SearchPositionRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response with search results.",
            "schema": {
              "$ref": "#/definitions/SearchPositionResponse"
            }
          }
        }
      }
    },
    "/api/Position/closeContract": {
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Closes a contract position based on the provided request.",
        "operationId": "Position_CloseContractPosition",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing the account ID and contract ID to close.",
            "schema": {
              "$ref": "#/definitions/CloseContractPositionRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response indicating the success or failure of the close operation.",
            "schema": {
              "$ref": "#/definitions/ClosePositionResponse"
            }
          }
        }
      }
    },
    "/api/Position/partialCloseContract": {
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Partially closes a contract position based on the provided request.",
        "operationId": "Position_PartialCloseContractPosition",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "description": "A request containing the account ID, contract ID, and size to close.",
            "schema": {
              "$ref": "#/definitions/PartialCloseContractPositionRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "A response indicating the success or failure of the partial close operation.",
            "schema": {
              "$ref": "#/definitions/PartialClosePositionResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SearchPositionResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/SearchPositionErrorCode"
        },
        "errorMessage": {
          "type": "string"
        },
        "positions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PositionModel"
          }
        }
      }
    },
    "SearchPositionErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound",
      "x-enumNames": [
        "Success",
        "AccountNotFound"
      ],
      "enum": [
        0,
        1
      ]
    },
    "PositionModel": {
      "type": "object",
      "required": [
        "id",
        "accountId",
        "contractId",
        "creationTimestamp",
        "type",
        "size",
        "averagePrice"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "contractId": {
          "type": "string"
        },
        "creationTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "type": {
          "$ref": "#/definitions/PositionType"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "averagePrice": {
          "type": "number",
          "format": "decimal"
        }
      }
    },
    "PositionType": {
      "type": "integer",
      "description": "0 = Undefined\n1 = Long\n2 = Short",
      "x-enumNames": [
        "Undefined",
        "Long",
        "Short"
      ],
      "enum": [
        0,
        1,
        2
      ]
    },
    "SearchPositionRequest": {
      "type": "object",
      "required": [
        "accountId"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ClosePositionResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/ClosePositionErrorCode"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "ClosePositionErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound\n2 = PositionNotFound\n3 = ContractNotFound\n4 = ContractNotActive\n5 = OrderRejected\n6 = OrderPending\n7 = UnknownError\n8 = AccountRejected",
      "x-enumNames": [
        "Success",
        "AccountNotFound",
        "PositionNotFound",
        "ContractNotFound",
        "ContractNotActive",
        "OrderRejected",
        "OrderPending",
        "UnknownError",
        "AccountRejected"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8
      ]
    },
    "CloseContractPositionRequest": {
      "type": "object",
      "required": [
        "accountId",
        "contractId"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "contractId": {
          "type": "string"
        }
      }
    },
    "PartialClosePositionResponse": {
      "type": "object",
      "required": [
        "success",
        "errorCode"
      ],
      "properties": {
        "success": {
          "type": "boolean"
        },
        "errorCode": {
          "$ref": "#/definitions/PartialClosePositionErrorCode"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "PartialClosePositionErrorCode": {
      "type": "integer",
      "description": "0 = Success\n1 = AccountNotFound\n2 = PositionNotFound\n3 = ContractNotFound\n4 = ContractNotActive\n5 = InvalidCloseSize\n6 = OrderRejected\n7 = OrderPending\n8 = UnknownError\n9 = AccountRejected",
      "x-enumNames": [
        "Success",
        "AccountNotFound",
        "PositionNotFound",
        "ContractNotFound",
        "ContractNotActive",
        "InvalidCloseSize",
        "OrderRejected",
        "OrderPending",
        "UnknownError",
        "AccountRejected"
      ],
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6,
        7,
        8,
        9
      ]
    },
    "PartialCloseContractPositionRequest": {
      "type": "object",
      "required": [
        "accountId",
        "contractId",
        "size"
      ],
      "properties": {
        "accountId": {
          "type": "integer",
          "format": "int32"
        },
        "contractId": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}